+++ src/lib/libhtml5.js
@@ -2374,6 +2374,12 @@ var LibraryHTML5 = {
     return requestAnimationFrame(tick);
   },
 
+  emscripten_queue_microtask: (cb, userData) => {
+    queueMicrotask(() => {
+      {{{ makeDynCall('vp', 'cb') }}}(userData);
+    });
+  },
+
   emscripten_get_device_pixel_ratio__proxy: 'sync',
   emscripten_get_device_pixel_ratio: () => {
 #if ENVIRONMENT_MAY_BE_NODE || ENVIRONMENT_MAY_BE_SHELL